requirements.txt
google-generativeai==0.3.2
argparse==1.4.0
python-dotenv==1.0.0
